Migrate from Jekyll to Nikola
Big update time! The original content in this repo was getting a bit
hard to maintain, and keep track of. Having to update every page when
adding a menu link, not keeping up to date with the newest information.
It was just not working out terribly well and fell into disuse.

I've seen found a new interest in Nikola. Already written enough about
this over in the actual blog posts, so I won't repeat it here. The gist
of it though -- this seems like it will be a *much* better solution for
us. Fingers crossed everything works online. And looking forward to
continuing to expand upon this in time.