Disable jekyll, verify, rebuild
Not entirely needed, as we do not use any filenames Jekyll would
consider special. But add it anyway to prevent potential problems. Add a
verification file so I can check the site indexing status. As part of
that process, regenerate parts of the site, which in turn updates some
metadata.

Effectively, no user-facing changes, this should bring.