Write new post, include tag/category descriptions
Spent quite a bit of time over the weekend getting this set up. Proper
descriptions and meta information for the tags and categories. Finished
these up yesterday. However, because of being in the middle of writing
a large post that was not ready, I held off on the commit.

The new post is finished, all tag metadata has been double checked. We
are ready to deploy and go live with our changes. Always gets me a
little nervous. You know, just kind of rebuild the entire website, no
big deal.

Either way -- enjoy! I know I'll be enjoying my lunch in a minute now
that I can call this finished. ;)