Correct typos
I was going through some `Vim` configuration options a bit ago, and came
across the spellcheck feature. Gave it a whirl and started noticing some
typos in posts here. Thankfully, I only have a handful of posts so
correction did not take a huge amount of time.

Probably a wise idea to enable spell check for `rST` files, or make a
habit of running it myself before publishing posts or other content.